feat: 初步完成基本功能

+++ src/main/java/com/s3/ftp/s3/S3FtpFile.java
@@ -2,19 +2,18 @@ package com.s3.ftp.s3;
 
 import org.apache.ftpserver.ftplet.FtpFile;
 import org.apache.ftpserver.ftplet.User;
+import org.apache.ftpserver.usermanager.impl.WriteRequest;
 import software.amazon.awssdk.awscore.exception.AwsServiceException;
 import software.amazon.awssdk.core.exception.SdkClientException;
 import software.amazon.awssdk.core.sync.RequestBody;
 import software.amazon.awssdk.services.s3.S3Client;
-import software.amazon.awssdk.services.s3.model.HeadObjectResponse;
-import software.amazon.awssdk.services.s3.model.NoSuchKeyException;
-import software.amazon.awssdk.services.s3.model.S3Object;
+import software.amazon.awssdk.services.s3.model.*;
 
-import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Stream;
 
 public class S3FtpFile implements FtpFile {
 
@@ -32,8 +31,6 @@ public class S3FtpFile implements FtpFile {
 
     private Long size;
 
-    private String owner;
-
     public S3FtpFile(S3Client client, String bucket, String key, User user) {
         this.client = client;
         this.bucket = bucket;
@@ -48,8 +45,14 @@ public class S3FtpFile implements FtpFile {
 
     @Override
     public String getName() {
-        int pos = key.lastIndexOf("/");
-        return pos != -1 ? key.substring(pos + 1) : key;
+        if (isDirectory()) {
+            int len = key.length();
+            int pos = key.lastIndexOf('/', len - 2);
+            return pos != -1 ? key.substring(pos + 1, len - 1) : key.substring(0, len - 1);
+        } else {
+            int pos = key.lastIndexOf("/");
+            return pos != -1 ? key.substring(pos + 1) : key;
+        }
     }
 
     @Override
@@ -59,19 +62,21 @@ public class S3FtpFile implements FtpFile {
 
     @Override
     public boolean isDirectory() {
-        return false;
+        return key.endsWith("/");
     }
 
     @Override
     public boolean isFile() {
-        return true;
+        return !isDirectory();
     }
 
     @Override
     public boolean doesExist() {
         if (exist == null) {
             try {
-                this.exist = getFileHead().isPresent();
+                this.exist = isFile()
+                        ? requestObjectHead().isPresent()
+                        : !client.listObjectsV2(req -> req.bucket(bucket).prefix(key)).contents().isEmpty();
             } catch (NoSuchKeyException e) {
                 this.exist = false;
             }
@@ -87,11 +92,24 @@ public class S3FtpFile implements FtpFile {
 
     @Override
     public boolean isWritable() {
-        return true;
+        return user.authorize(new WriteRequest(getPhysicalFile().toString())) != null;
     }
 
     @Override
     public boolean isRemovable() {
+        if (key.equals("/")) {
+            return false;
+        }
+
+        if (!isWritable()) {
+            return false;
+        }
+
+        if (isDirectory()) {
+            return listFiles().stream()
+                    .allMatch(v -> v.getAbsolutePath().equals(getAbsolutePath()));
+        }
+
         return true;
     }
 
@@ -113,7 +131,7 @@ public class S3FtpFile implements FtpFile {
     @Override
     public long getLastModified() {
         if (lastModified == null) {
-            this.lastModified = getFileHead()
+            this.lastModified = requestObjectHead()
                     .map(v -> v.lastModified().getEpochSecond())
                     .orElse(null);
         }
@@ -128,8 +146,12 @@ public class S3FtpFile implements FtpFile {
 
     @Override
     public long getSize() {
+        if (isDirectory()) {
+            return 0;
+        }
+
         if (size == null) {
-            this.size = getFileHead().map(HeadObjectResponse::contentLength).orElse(null);
+            this.size = requestObjectHead().map(HeadObjectResponse::contentLength).orElse(null);
         }
 
         return size == null ? -1 : size;
@@ -142,8 +164,13 @@ public class S3FtpFile implements FtpFile {
 
     @Override
     public boolean mkdir() {
+        String dirKey = key.endsWith("/") ? key : key + "/";
+
         try {
-            client.putObject(req -> req.bucket(bucket).key(key + "/"), RequestBody.empty());
+            client.putObject(
+                    req -> req.bucket(bucket).key(dirKey).contentType("application/x-directory"),
+                    RequestBody.empty()
+            );
             return true;
         } catch (AwsServiceException | SdkClientException e) {
             return false;
@@ -152,6 +179,10 @@ public class S3FtpFile implements FtpFile {
 
     @Override
     public boolean delete() {
+        if (!isRemovable()) {
+            return false;
+        }
+
         try {
             client.deleteObject(req -> req.bucket(bucket).key(key));
             return true;
@@ -162,13 +193,21 @@ public class S3FtpFile implements FtpFile {
 
     @Override
     public boolean move(FtpFile destination) {
+        if (!destination.isWritable() || !isWritable()) {
+            return false;
+        }
+
         try {
+            // 复制数据
             client.copyObject(req -> req
                     .sourceBucket(bucket)
                     .sourceKey(key)
                     .destinationBucket(bucket)
                     .destinationKey(destination.getAbsolutePath())
             );
+            // 确认数据复制成功
+            client.headObject(req -> req.bucket(bucket).key(destination.getAbsolutePath()));
+            // 删除源数据
             client.deleteObject(req -> req.bucket(bucket).key(key));
             return true;
         } catch (AwsServiceException | SdkClientException e) {
@@ -178,15 +217,30 @@ public class S3FtpFile implements FtpFile {
 
     @Override
     public List<? extends FtpFile> listFiles() {
-        return client.listObjectsV2(req -> req.bucket(bucket).prefix(key + "/")).contents()
-                .stream()
-                .map(S3Object::key)
+        if (!isDirectory()) {
+            return null;
+        }
+
+        ListObjectsV2Request.Builder builder = ListObjectsV2Request.builder()
+                .bucket(bucket)
+                //TODO configuration
+                .maxKeys(10000)
+                .delimiter("/");
+        if (!key.equals("/")) {
+            builder.prefix(key);
+        }
+        ListObjectsV2Response resp = client.listObjectsV2(builder.build());
+
+        return Stream.concat(
+                        resp.commonPrefixes().stream().map(CommonPrefix::prefix),
+                        resp.contents().stream().map(S3Object::key)
+                )
                 .map(key -> new S3FtpFile(client, bucket, key, user))
                 .toList();
     }
 
     @Override
-    public OutputStream createOutputStream(long offset) throws IOException {
+    public OutputStream createOutputStream(long offset) {
         return new S3OutputStream(client, bucket, key, offset);
     }
 
@@ -195,7 +249,7 @@ public class S3FtpFile implements FtpFile {
         return client.getObject(req -> req.bucket(bucket).key(key).range("%s-".formatted(offset)));
     }
 
-    private Optional<HeadObjectResponse> getFileHead() {
+    private Optional<HeadObjectResponse> requestObjectHead() {
         try {
             return Optional.of(client.headObject(req -> req.bucket(bucket).key(key)));
         } catch (AwsServiceException | SdkClientException e) {
